Converted to library module of core features to release as .aar file
