+++ build.gradle
@@ -1,4 +1,5 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
 id 'com.android.application' version '8.1.0-rc01' apply false
+    id 'com.android.library' version '8.1.0-rc01' apply false
 }
