+++ emulatordetection/.gitignore
