+++ emulatordetection/proguard-rules.pro
