+++ emulatordetection/src/main/java/com/reveny/emulator/detection/EmulatorDetection.java
@@ -1,0 +1,13 @@
+package com.reveny.emulator.detection;
+
+public class EmulatorDetection {
+
+    public EmulatorDetection()
+    {
+        System.loadLibrary("emulatordetector");
+    }
+
+    public native boolean isDetected();
+    public native String getResult();
+
+}
