+++ emulatordetection/src/main/jni/Android.mk
