+++ emulatordetection/src/main/jni/Application.mk
