+++ emulatordetection/src/main/jni/Include/EmulatorData.h
