+++ emulatordetection/src/main/jni/Include/FileHelper.h
