+++ emulatordetection/src/main/jni/Include/StringUtility.h
