+++ sample/src/main/res/drawable/circle_check_solid.xml
