+++ sample/src/main/res/drawable/xmark.xml
