+++ settings.gradle
@@ -14,4 +14,5 @@ dependencyResolutionManagement {
 }
 
 rootProject.name = "AndroidEmulatorDetection"
-include ':app'
+include ':emulatordetection'
+include ':sample'
